+++ dantecely/server-recipe/index.js
@@ -1,37 +1,11 @@
 require("dotenv").config();
 const express = require("express");
+const router = require("./router");
 
 const app = express();
 const { PORT } = process.env;
 
-const recipe = {
-  title: "Delicious Pasta Carbonara",
-  peopleAmount: {
-    max: 4,
-    min: 2,
-  },
-  time: 30, // en minutos
-  cost: "$$", // Escala de costo: $ (barato), $$ (moderado), $$$ (caro)
-  level: "Fácil", // Niveles: Fácil, Intermedio, Difícil
-  steps: [
-    "Hierve agua y agrega sal.",
-    "Añade la pasta y cocina hasta que esté al dente.",
-    "En una sartén, saltea el tocino hasta que esté crujiente.",
-    "Agrega ajo picado y cocina hasta que esté fragante.",
-    "Bate los huevos con queso parmesano en un tazón aparte.",
-    "Escurre la pasta y agrégala a la sartén con el tocino.",
-    "Agrega la mezcla de huevo y queso, revuelve rápidamente para evitar que los huevos se cuajen demasiado.",
-    "Añade pimienta negra molida al gusto.",
-    "Sirve caliente y disfruta de tu deliciosa Pasta Carbonara.",
-  ],
-};
-
-// middleware
-const callback = (req, res) => {
-  res.send(recipe);
-};
-
-app.get("/", callback);
+app.use(router);
 
 app.listen(PORT, () => {
   console.log(`Server is running on port ${PORT}`);
