+++ dantecely/server-recipe/router/recipe/delete.js
@@ -1,0 +1,7 @@
+const removeRecipes = (req, res) => {
+  // Operaciones con la base de datos
+
+  res.send("Eliminé una receta");
+};
+
+module.exports = { removeRecipes };
