+++ dantecely/server-recipe/router/recipe/post.js
@@ -1,0 +1,7 @@
+const updateRecipe = (req, res) => {
+  // Operaciones con la base de datos
+
+  res.send("Actualizar una receta");
+};
+
+module.exports = { updateRecipe };
