+++ dantecely/server-recipe/router/recipe/put.js
@@ -1,0 +1,7 @@
+const addRecipe = (req, res) => {
+  // Operaciones con la base de datos
+
+  res.send("Agregué una receta");
+};
+
+module.exports = { addRecipe };
